+++ src/Contracts/SpreadsheetGeneratorServiceInterface.php
@@ -9,6 +9,8 @@ use Symfony\Component\HttpFoundation\StreamedResponse;
 interface SpreadsheetGeneratorServiceInterface
 {
     public const WRITER = 'writer';
+    public const XLSX_WRITER = 'xlsx';
+    public const CSV_WRITER = 'csv';
     public const SHEETS = 'sheets';
     public const CONTENT_FILENAME = 'filename';
     public const CONTENT_DISPOSITION = 'disposition';
