+++ src/Resources/config/services.xml
@@ -17,7 +17,6 @@
         <service id="EMS\CommonBundle\Contracts\CoreApi\CoreApiFactoryInterface" alias="ems_common.core_api.factory" />
         <service id="EMS\CommonBundle\Contracts\ExpressionServiceInterface" alias="ems_common.service.expression_service" />
         <service id="EMS\CommonBundle\Contracts\SpreadsheetGeneratorServiceInterface" alias="ems_common.service.spreadsheet_generator_service" />
-        <service id="EMS\CommonBundle\Contracts\CsvGeneratorServiceInterface" alias="ems_common.service.csv_generator_service" />
 
         <service id="ems_common.core_api.factory" class="EMS\CommonBundle\Common\CoreApi\CoreApiFactory">
             <argument type="service" id="logger" />
@@ -61,9 +60,6 @@
             <argument type="service" id="logger" />
         </service>
         <service id="ems_common.service.spreadsheet_generator_service" class="EMS\CommonBundle\Common\SpreadsheetGeneratorService">
-            <argument type="service" id="logger" />
-        </service>
-        <service id="ems_common.service.csv_generator_service" class="EMS\CommonBundle\Common\CsvGeneratorService">
             <argument type="service" id="logger" />
         </service>
         <service id="EMS\CommonBundle\Service\ElasticaService" alias="ems_common.service.elastica"/>
