+++ tests/Unit/Common/SpreadsheetGeneratorTest.php
@@ -4,6 +4,7 @@ namespace EMS\CommonBundle\Tests\Unit\Common;
 
 use EMS\CommonBundle\Common\SpreadsheetGeneratorService;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class SpreadsheetGeneratorTest extends TestCase
 {
@@ -16,11 +17,31 @@ class SpreadsheetGeneratorTest extends TestCase
         parent::setUp();
     }
 
-    public function testConfigToSpreadsheet(): void
+    public function testConfigToExcel(): void
     {
         $config = \json_decode('{"filename":"export","writer":"xlsx","active_sheet":0,"sheets":[{"name":"Export form","color":"#FF0000","rows":[["apple","banana"],["pineapple","strawberry"]]},{"name":"Export form sheet 2","rows":[["a1","a2"],["b1","b3"]]}]}', true);
         $this->assertSame('Export form', $this->callMethod($this->spreadSheetGenerator, 'buildUpSheets', [$config])->getActiveSheet()->getTitle());
         $this->assertSame('pineapple', $this->callMethod($this->spreadSheetGenerator, 'buildUpSheets', [$config])->getActiveSheet()->getCell('A2')->getValue());
+    }
+
+    public function testConfigToCsv(): void
+    {
+        $config = \json_decode('{"filename":"test-export", "writer": "csv","disposition":"inline","sheets":[[["apple","banana"],["pineapple","strawberry"],["àï$@,& & \\" \' ! @ # $ €", "foobar"]]]}', true);
+        /** @var StreamedResponse $csv */
+        $csv = $this->callMethod($this->spreadSheetGenerator, 'getCsvResponse', [$config]);
+
+        //https://github.com/symfony/symfony/issues/25005
+        \ob_start();
+        $csv->send();
+        $getContent = \ob_get_contents();
+        \ob_end_clean();
+
+        $lines = \explode(PHP_EOL, $getContent);
+        $this->assertCount(4, $lines);
+        $this->assertSame('apple,banana', $lines[0]);
+        $this->assertSame('pineapple,strawberry', $lines[1]);
+        $this->assertSame('"àï$@,& & "" \' ! @ # $ €",foobar', $lines[2]);
+        $this->assertSame('', $lines[3]);
     }
 
     /**
